Converted to Rmd file & begin completion
I misread the instruction that week and thought that completing the assignment was to just load it into Github, not to do the coding in R since the file was not in a Rmd format. With more knowledge from the class and after reviewing my files, I am now going back to makeup for the work.